Add autogenerated autopilot briefing for Economic Power demo
