+++ demo/Economic-Power-v0/reports/owner-autopilot-brief.md
@@ -1,0 +1,82 @@
+# Economic Power Autopilot Brief
+
+Scenario: AGIJobs Economic Power Launch (economic-power-v0-baseline)
+Mission: Sustain unstoppable economic acceleration with deterministic owner oversight.
+
+Cadence: 6.7h • Dominance 99.4% (total-dominion)
+Control coverage: 100.0% — Owner multi-sig holds deterministic runbooks for every critical surface.
+Safety mesh score: 97.9% • Custody: 100.0%
+Owner multi-sig exerts absolute dominion – every command, pause, and upgrade is scripted and rehearsed.
+
+## Guardrails
+- Pause • npm run owner:system-pause
+- Resume • npm run owner:update-all
+- validatorConfidence < 0.95 → npm run owner:system-pause
+- automationScore < 0.8 → npm run owner:parameters
+- treasuryAfterRun < 1800000 → npm run owner:audit
+- Control drills 100.0% (total-control)
+
+## Command sequence
+1. [JOBS] npm run owner:program -- --program job-ai-lab-fusion — Dispatch Helios and Atlas to fuse AI labs into a consolidated planetary accelerator.
+2. [JOBS] npm run owner:program -- --program job-supply-chain — Reconfigure autonomous supply mesh coverage for instant cross-border fulfilment.
+3. [VALIDATORS] npm run owner:program -- --program validator-alpha — Escalate Validator Alpha quorum weighting and deploy fresh attestations.
+4. [VALIDATORS] npm run owner:program -- --program validator-beta — Rotate Validator Beta committees and enforce staking deltas.
+5. [STABLECOINADAPTERS] npm run owner:program -- --program adapter-usdc — Upgrade the USDC bridge adapter with deterministic slippage guards.
+6. [MODULES] npm run owner:program -- --program module-job-registry — Promote the JobRegistry v2 deployment with zero-downtime migration.
+7. [MODULES] npm run owner:program -- --program module-stake-manager — Adjust stake thresholds and payout cadence for StakeManager.
+8. [TREASURY] npm run owner:program -- --program treasury-liquidity — Rebalance AGI/USDC buffers and top up validator incentive pools.
+9. [TREASURY] npm run owner:program -- --program treasury-yield — Deploy surplus treasury capital into protocol-aligned yield strategies.
+10. [ORCHESTRATOR] npm run owner:program -- --program orchestrator-rebalance — Re-weight autonomous routing priorities to maximise throughput.
+11. [ORCHESTRATOR] npm run owner:program -- --program orchestrator-pause-drill — Run full-stack pause/resume drill to validate emergency readiness.
+
+## Safety mesh readiness
+- Pause command: `npm run owner:system-pause`
+- Resume command: `npm run owner:update-all`
+- Response cadence: 9m (target ≤ 15m)
+- Mesh status: pause ready, resume ready
+- Safety score: 97.9% • Command coverage: 100.0%
+
+## Telemetry checkpoints
+- Economic dominance index: 99.4%
+- Superintelligence index: 96.8%
+- Capital velocity: 26290.99 AGI/h
+- Global expansion readiness: 96.6%
+- Shock resilience: 98.0%
+
+## Shock resilience posture
+- Score: 98.0% (impregnable)
+- Summary: Shockwaves are fully absorbed – treasury buffers, guardrails, and automation form an impregnable shield.
+- Drivers:
+  - Stability index 76.1%
+  - Guardrail coverage 98.7%
+  - Risk dampening 73.4%
+  - Emergency depth 3 contact(s)
+  - Alert mesh 2 channel(s)
+  - Operations buffer 14.4% of treasury
+  - Automation density 85.2%
+- Recommended actions:
+  - Maintain drill cadence and telemetry mirroring to preserve impregnable resilience.
+
+## Dominance signals
+- Command coverage 100.0%
+- Safety mesh 97.9%
+- Custody 100.0%
+- Guardrails 6
+- Response 9m
+- Control drills 100.0% (total-control)
+- Shock resilience 98.0% (impregnable)
+
+## Coverage detail
+- Job orchestration: 100.0%
+- Validator sovereignty: 100.0%
+- Stablecoin adapters: 100.0%
+- Protocol modules: 100.0%
+- Parameter overrides: 100.0%
+- Pause command: 100.0%
+- Resume command: 100.0%
+- Treasury programs: 100.0%
+- Orchestrator mesh: 100.0%
+
+## Recommended actions
+- Maintain autopilot cadence and periodic drills to preserve total dominion.
+
